Move cluster config source to its own file
